refactor(tests): re-pin platform devices and ui-base test paths to subjects
